tarot-ai-generator/requirements.txt
langchain-community>=0.0.20
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
zhipuai>=2.0.1
openai>=1.0.0
ollama>=0.1.0
PyYAML>=6.0
